refactor(mcp): extract single-writer election into single_writer.rs
lib.rs mixed the 14-tool MCP surface with the entire single-writer
election + promoter subsystem. Move election/promotion (acquire_serve_lock,
run_server, promoter_task, validate_pinned_state, signal waiters, and the
PromoterArgs/ServeLockOutcome types) into single_writer.rs over
`use super::*`; lib.rs re-exports the public entry points.

Pure code movement, no API change (92 tests pass unchanged). lib.rs drops
to 3602 lines. Items the in-lib test module still references are pub(crate);
the test-only call counter moved to EOF to satisfy items-after-test-module.
